Network.AWS.Glue.BatchDeleteConnection

Description

Deletes a list of connection definitions from the Data Catalog.

Creating a Request

batchDeleteConnection :: BatchDeleteConnection

Creates a value of BatchDeleteConnection with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bdcCatalogId - The ID of the Data Catalog in which the connections reside. If none is supplied, the AWS account ID is used by default.

bdcConnectionNameList - A list of names of the connections to delete.

data BatchDeleteConnection

See: batchDeleteConnection smart constructor.

Request Lenses

bdcCatalogId :: Lens' BatchDeleteConnection (Maybe Text)

The ID of the Data Catalog in which the connections reside. If none is supplied, the AWS account ID is used by default.

bdcConnectionNameList :: Lens' BatchDeleteConnection [Text]

A list of names of the connections to delete.

Destructuring the Response

batchDeleteConnectionResponse

Arguments

:: Int
bdcrsResponseStatus

-> BatchDeleteConnectionResponse

Creates a value of BatchDeleteConnectionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bdcrsSucceeded - A list of names of the connection definitions that were successfully deleted.

bdcrsErrors - A map of the names of connections that were not successfully deleted to error details.

bdcrsResponseStatus - -- | The response status code.

data BatchDeleteConnectionResponse

See: batchDeleteConnectionResponse smart constructor.

Response Lenses

bdcrsSucceeded :: Lens' BatchDeleteConnectionResponse [Text]

A list of names of the connection definitions that were successfully deleted.

bdcrsErrors :: Lens' BatchDeleteConnectionResponse (HashMap Text ErrorDetail)

A map of the names of connections that were not successfully deleted to error details.

bdcrsResponseStatus :: Lens' BatchDeleteConnectionResponse Int

- | The response status code.