Network.AWS.Glue.BatchDeleteTable

Description

Deletes multiple tables at once.

Creating a Request

batchDeleteTable

Arguments

:: Text
bdtDatabaseName

-> BatchDeleteTable

Creates a value of BatchDeleteTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bdtCatalogId - The ID of the Data Catalog where the table resides. If none is supplied, the AWS account ID is used by default.

bdtDatabaseName - The name of the catalog database where the tables to delete reside.

bdtTablesToDelete - A list of the table to delete.

data BatchDeleteTable

See: batchDeleteTable smart constructor.

Request Lenses

bdtCatalogId :: Lens' BatchDeleteTable (Maybe Text)

The ID of the Data Catalog where the table resides. If none is supplied, the AWS account ID is used by default.

bdtDatabaseName :: Lens' BatchDeleteTable Text

The name of the catalog database where the tables to delete reside.

bdtTablesToDelete :: Lens' BatchDeleteTable [Text]

A list of the table to delete.

Destructuring the Response

batchDeleteTableResponse

Arguments

:: Int
bdtrsResponseStatus

-> BatchDeleteTableResponse

Creates a value of BatchDeleteTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bdtrsErrors - A list of errors encountered in attempting to delete the specified tables.

bdtrsResponseStatus - -- | The response status code.

data BatchDeleteTableResponse

See: batchDeleteTableResponse smart constructor.

Response Lenses

bdtrsErrors :: Lens' BatchDeleteTableResponse [TableError]

A list of errors encountered in attempting to delete the specified tables.

bdtrsResponseStatus :: Lens' BatchDeleteTableResponse Int

- | The response status code.