Network.AWS.Glue.BatchGetPartition

Description

Retrieves partitions in a batch request.

Creating a Request

batchGetPartition

Arguments

:: Text
bgpDatabaseName

-> Text
bgpTableName

-> BatchGetPartition

Creates a value of BatchGetPartition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgpCatalogId - The ID of the Data Catalog where the partitions in question reside. If none is supplied, the AWS account ID is used by default.

bgpDatabaseName - The name of the catalog database where the partitions reside.

bgpTableName - The name of the partitions' table.

bgpPartitionsToGet - A list of partition values identifying the partitions to retrieve.

data BatchGetPartition

See: batchGetPartition smart constructor.

Request Lenses

bgpCatalogId :: Lens' BatchGetPartition (Maybe Text)

The ID of the Data Catalog where the partitions in question reside. If none is supplied, the AWS account ID is used by default.

bgpDatabaseName :: Lens' BatchGetPartition Text

The name of the catalog database where the partitions reside.

bgpTableName :: Lens' BatchGetPartition Text

The name of the partitions' table.

bgpPartitionsToGet :: Lens' BatchGetPartition [PartitionValueList]

A list of partition values identifying the partitions to retrieve.

Destructuring the Response

batchGetPartitionResponse

Arguments

:: Int
bgprsResponseStatus

-> BatchGetPartitionResponse

Creates a value of BatchGetPartitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

bgprsUnprocessedKeys - A list of the partition values in the request for which partions were not returned.

bgprsPartitions - A list of the requested partitions.

bgprsResponseStatus - -- | The response status code.

data BatchGetPartitionResponse

See: batchGetPartitionResponse smart constructor.

Response Lenses

bgprsUnprocessedKeys :: Lens' BatchGetPartitionResponse [PartitionValueList]

A list of the partition values in the request for which partions were not returned.

bgprsPartitions :: Lens' BatchGetPartitionResponse [Partition]

A list of the requested partitions.

bgprsResponseStatus :: Lens' BatchGetPartitionResponse Int

- | The response status code.