Network.AWS.Glue.CreateCrawler

Description

Creates a new Crawler with specified targets, role, configuration, and optional schedule. At least one crawl target must be specified, in either the s3Targets or the jdbcTargets field.

Creating a Request

createCrawler

Arguments

:: Text
ccName

-> Text
ccRole

-> Text
ccDatabaseName

-> CrawlerTargets
ccTargets

-> CreateCrawler

Creates a value of CreateCrawler with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ccSchemaChangePolicy - Policy for the crawler's update and deletion behavior.

ccSchedule - A cron expression that can be used as a Cloudwatch event (see CloudWatch Schedule Expression Syntax . For example, to run every day at 12:15 UTC, specify: cron(15 12 * * ? *) .

ccClassifiers - A list of custom Classifier names that the user has registered. By default, all AWS classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

ccTablePrefix - The table prefix used for catalog tables created.

ccDescription - A description of the new Crawler .

ccName - Name of the new Crawler .

ccRole - The AWS ARN of the IAM role used by the new Crawler to access customer resources.

ccDatabaseName - The Glue Database where results will be stored, such as: arn:aws:daylight:us-east-1::databasesometable* .

ccTargets - A list of collection of targets to crawl.

data CreateCrawler

See: createCrawler smart constructor.

Request Lenses

ccSchemaChangePolicy :: Lens' CreateCrawler (Maybe SchemaChangePolicy)

Policy for the crawler's update and deletion behavior.

ccSchedule :: Lens' CreateCrawler (Maybe Text)

A cron expression that can be used as a Cloudwatch event (see CloudWatch Schedule Expression Syntax . For example, to run every day at 12:15 UTC, specify: cron(15 12 * * ? *) .

ccClassifiers :: Lens' CreateCrawler [Text]

A list of custom Classifier names that the user has registered. By default, all AWS classifiers are included in a crawl, but these custom classifiers always override the default classifiers for a given classification.

ccTablePrefix :: Lens' CreateCrawler (Maybe Text)

The table prefix used for catalog tables created.

ccDescription :: Lens' CreateCrawler (Maybe Text)

A description of the new Crawler .

ccName :: Lens' CreateCrawler Text

Name of the new Crawler .

ccRole :: Lens' CreateCrawler Text

The AWS ARN of the IAM role used by the new Crawler to access customer resources.

ccDatabaseName :: Lens' CreateCrawler Text

The Glue Database where results will be stored, such as: arn:aws:daylight:us-east-1::databasesometable* .

ccTargets :: Lens' CreateCrawler CrawlerTargets

A list of collection of targets to crawl.

Destructuring the Response

createCrawlerResponse

Arguments

:: Int
crersResponseStatus

-> CreateCrawlerResponse

Creates a value of CreateCrawlerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crersResponseStatus - -- | The response status code.

data CreateCrawlerResponse

See: createCrawlerResponse smart constructor.

Response Lenses

crersResponseStatus :: Lens' CreateCrawlerResponse Int

- | The response status code.