Network.AWS.Glue.CreateDevEndpoint

Description

Creates a new DevEndpoint.

Creating a Request

createDevEndpoint

Arguments

:: Text
cdeEndpointName

-> Text
cdeRoleARN

-> Text
cdeSubnetId

-> CreateDevEndpoint

Creates a value of CreateDevEndpoint with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdeExtraPythonLibsS3Path - Path to one or more Python libraries in an S3 bucket that should be loaded in your DevEndpoint.

cdePublicKey - The public key to use for authentication.

cdeNumberOfNodes - The number of nodes to use.

cdeExtraJARsS3Path - Path to one or more Java Jars in an S3 bucket that should be loaded in your DevEndpoint.

cdeEndpointName - The name to be assigned to the new DevEndpoint.

cdeRoleARN - The IAM role for the DevEndpoint.

cdeSecurityGroupIds - Security group IDs for the security groups to be used by the new DevEndpoint.

cdeSubnetId - The subnet ID for the new DevEndpoint to use.

data CreateDevEndpoint

See: createDevEndpoint smart constructor.

Request Lenses

cdeExtraPythonLibsS3Path :: Lens' CreateDevEndpoint (Maybe Text)

Path to one or more Python libraries in an S3 bucket that should be loaded in your DevEndpoint.

cdePublicKey :: Lens' CreateDevEndpoint (Maybe Text)

The public key to use for authentication.

cdeNumberOfNodes :: Lens' CreateDevEndpoint (Maybe Int)

The number of nodes to use.

cdeExtraJARsS3Path :: Lens' CreateDevEndpoint (Maybe Text)

Path to one or more Java Jars in an S3 bucket that should be loaded in your DevEndpoint.

cdeEndpointName :: Lens' CreateDevEndpoint Text

The name to be assigned to the new DevEndpoint.

cdeRoleARN :: Lens' CreateDevEndpoint Text

The IAM role for the DevEndpoint.

cdeSecurityGroupIds :: Lens' CreateDevEndpoint [Text]

Security group IDs for the security groups to be used by the new DevEndpoint.

cdeSubnetId :: Lens' CreateDevEndpoint Text

The subnet ID for the new DevEndpoint to use.

Destructuring the Response

createDevEndpointResponse

Arguments

:: Int
cdersResponseStatus

-> CreateDevEndpointResponse

Creates a value of CreateDevEndpointResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cdersStatus - The current status of the new DevEndpoint.

cdersFailureReason - The reason for a current failure in this DevEndpoint.

cdersEndpointName - The name assigned to the new DevEndpoint.

cdersExtraPythonLibsS3Path - Path to one or more Python libraries in an S3 bucket that will be loaded in your DevEndpoint.

cdersSecurityGroupIds - The security groups assigned to the new DevEndpoint.

cdersVPCId - The ID of the VPC used by this DevEndpoint.

cdersSubnetId - The subnet ID assigned to the new DevEndpoint.

cdersNumberOfNodes - The number of nodes in this DevEndpoint.

cdersAvailabilityZone - The AWS availability zone where this DevEndpoint is located.

cdersExtraJARsS3Path - Path to one or more Java Jars in an S3 bucket that will be loaded in your DevEndpoint.

cdersCreatedTimestamp - The point in time at which this DevEndpoint was created.

cdersYarnEndpointAddress - The address of the YARN endpoint used by this DevEndpoint.

cdersRoleARN - The AWS ARN of the role assigned to the new DevEndpoint.

cdersResponseStatus - -- | The response status code.

data CreateDevEndpointResponse

See: createDevEndpointResponse smart constructor.

Response Lenses

cdersStatus :: Lens' CreateDevEndpointResponse (Maybe Text)

The current status of the new DevEndpoint.

cdersFailureReason :: Lens' CreateDevEndpointResponse (Maybe Text)

The reason for a current failure in this DevEndpoint.

cdersEndpointName :: Lens' CreateDevEndpointResponse (Maybe Text)

The name assigned to the new DevEndpoint.

cdersExtraPythonLibsS3Path :: Lens' CreateDevEndpointResponse (Maybe Text)

Path to one or more Python libraries in an S3 bucket that will be loaded in your DevEndpoint.

cdersSecurityGroupIds :: Lens' CreateDevEndpointResponse [Text]

The security groups assigned to the new DevEndpoint.

cdersVPCId :: Lens' CreateDevEndpointResponse (Maybe Text)

The ID of the VPC used by this DevEndpoint.

cdersSubnetId :: Lens' CreateDevEndpointResponse (Maybe Text)

The subnet ID assigned to the new DevEndpoint.

cdersNumberOfNodes :: Lens' CreateDevEndpointResponse (Maybe Int)

The number of nodes in this DevEndpoint.

cdersAvailabilityZone :: Lens' CreateDevEndpointResponse (Maybe Text)

The AWS availability zone where this DevEndpoint is located.

cdersExtraJARsS3Path :: Lens' CreateDevEndpointResponse (Maybe Text)

Path to one or more Java Jars in an S3 bucket that will be loaded in your DevEndpoint.

cdersCreatedTimestamp :: Lens' CreateDevEndpointResponse (Maybe UTCTime)

The point in time at which this DevEndpoint was created.

cdersYarnEndpointAddress :: Lens' CreateDevEndpointResponse (Maybe Text)

The address of the YARN endpoint used by this DevEndpoint.

cdersRoleARN :: Lens' CreateDevEndpointResponse (Maybe Text)

The AWS ARN of the role assigned to the new DevEndpoint.

cdersResponseStatus :: Lens' CreateDevEndpointResponse Int

- | The response status code.