Network.AWS.Glue.CreateJob

Description

Creates a new job.

Creating a Request

createJob

Arguments

:: Text
cjName

-> Text
cjRole

-> JobCommand
cjCommand

-> CreateJob

Creates a value of CreateJob with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjConnections - The connections used for this job.

cjLogURI - Location of the logs for this job.

cjMaxRetries - The maximum number of times to retry this job if it fails.

cjExecutionProperty - An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

cjAllocatedCapacity - The number of capacity units allocated to this job.

cjDefaultArguments - The default parameters for this job.

cjDescription - Description of the job.

cjName - The name you assign to this job.

cjRole - The role associated with this job.

cjCommand - The JobCommand that executes this job.

data CreateJob

See: createJob smart constructor.

Request Lenses

cjConnections :: Lens' CreateJob (Maybe ConnectionsList)

The connections used for this job.

cjLogURI :: Lens' CreateJob (Maybe Text)

Location of the logs for this job.

cjMaxRetries :: Lens' CreateJob (Maybe Int)

The maximum number of times to retry this job if it fails.

cjExecutionProperty :: Lens' CreateJob (Maybe ExecutionProperty)

An ExecutionProperty specifying the maximum number of concurrent runs allowed for this job.

cjAllocatedCapacity :: Lens' CreateJob (Maybe Int)

The number of capacity units allocated to this job.

cjDefaultArguments :: Lens' CreateJob (HashMap Text Text)

The default parameters for this job.

cjDescription :: Lens' CreateJob (Maybe Text)

Description of the job.

cjName :: Lens' CreateJob Text

The name you assign to this job.

cjRole :: Lens' CreateJob Text

The role associated with this job.

cjCommand :: Lens' CreateJob JobCommand

The JobCommand that executes this job.

Destructuring the Response

createJobResponse

Arguments

:: Int
cjrsResponseStatus

-> CreateJobResponse

Creates a value of CreateJobResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cjrsName - The unique name of the new job that has been created.

cjrsResponseStatus - -- | The response status code.

data CreateJobResponse

See: createJobResponse smart constructor.

Response Lenses

cjrsName :: Lens' CreateJobResponse (Maybe Text)

The unique name of the new job that has been created.

cjrsResponseStatus :: Lens' CreateJobResponse Int

- | The response status code.