Network.AWS.Glue.CreateTable

Description

Creates a new table definition in the Data Catalog.

Creating a Request

createTable

Arguments

:: Text
ctDatabaseName

-> TableInput
ctTableInput

-> CreateTable

Creates a value of CreateTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

ctCatalogId - The ID of the Data Catalog in which to create the Table . If none is supplied, the AWS account ID is used by default.

ctDatabaseName - The catalog database in which to create the new table.

ctTableInput - The TableInput object that defines the metadata table to create in the catalog.

data CreateTable

See: createTable smart constructor.

Request Lenses

ctCatalogId :: Lens' CreateTable (Maybe Text)

The ID of the Data Catalog in which to create the Table . If none is supplied, the AWS account ID is used by default.

ctDatabaseName :: Lens' CreateTable Text

The catalog database in which to create the new table.

ctTableInput :: Lens' CreateTable TableInput

The TableInput object that defines the metadata table to create in the catalog.

Destructuring the Response

createTableResponse

Arguments

:: Int
cttrsResponseStatus

-> CreateTableResponse

Creates a value of CreateTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cttrsResponseStatus - -- | The response status code.

data CreateTableResponse

See: createTableResponse smart constructor.

Response Lenses

cttrsResponseStatus :: Lens' CreateTableResponse Int

- | The response status code.