Network.AWS.Glue.DeleteTrigger

Description

Deletes a specified trigger.

Creating a Request

deleteTrigger

Arguments

:: Text
dttName

-> DeleteTrigger

Creates a value of DeleteTrigger with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dttName - The name of the trigger to delete.

data DeleteTrigger

See: deleteTrigger smart constructor.

Request Lenses

dttName :: Lens' DeleteTrigger Text

The name of the trigger to delete.

Destructuring the Response

deleteTriggerResponse

Arguments

:: Int
delrsResponseStatus

-> DeleteTriggerResponse

Creates a value of DeleteTriggerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

delrsName - The name of the trigger that was deleted.

delrsResponseStatus - -- | The response status code.

data DeleteTriggerResponse

See: deleteTriggerResponse smart constructor.

Response Lenses

delrsName :: Lens' DeleteTriggerResponse (Maybe Text)

The name of the trigger that was deleted.

delrsResponseStatus :: Lens' DeleteTriggerResponse Int

- | The response status code.