Network.AWS.Glue.GetCrawlerMetrics

Description

Retrieves metrics about specified crawlers.

Creating a Request

getCrawlerMetrics :: GetCrawlerMetrics

Creates a value of GetCrawlerMetrics with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcmNextToken - A continuation token, if this is a continuation call.

gcmMaxResults - The maximum size of a list to return.

gcmCrawlerNameList - A list of the names of crawlers about which to retrieve metrics.

data GetCrawlerMetrics

See: getCrawlerMetrics smart constructor.

Request Lenses

gcmNextToken :: Lens' GetCrawlerMetrics (Maybe Text)

A continuation token, if this is a continuation call.

gcmMaxResults :: Lens' GetCrawlerMetrics (Maybe Natural)

The maximum size of a list to return.

gcmCrawlerNameList :: Lens' GetCrawlerMetrics [Text]

A list of the names of crawlers about which to retrieve metrics.

Destructuring the Response

getCrawlerMetricsResponse

Arguments

:: Int
gcmrsResponseStatus

-> GetCrawlerMetricsResponse

Creates a value of GetCrawlerMetricsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gcmrsCrawlerMetricsList - A list of metrics for the specified crawler.

gcmrsNextToken - A continuation token, if the returned list does not contain the last metric available.

gcmrsResponseStatus - -- | The response status code.

data GetCrawlerMetricsResponse

See: getCrawlerMetricsResponse smart constructor.

Response Lenses

gcmrsCrawlerMetricsList :: Lens' GetCrawlerMetricsResponse [CrawlerMetrics]

A list of metrics for the specified crawler.

gcmrsNextToken :: Lens' GetCrawlerMetricsResponse (Maybe Text)

A continuation token, if the returned list does not contain the last metric available.

gcmrsResponseStatus :: Lens' GetCrawlerMetricsResponse Int

- | The response status code.