Network.AWS.Glue.GetCrawlers

Description

Retrieves metadata for all Crawlers defined in the customer account.

Creating a Request

getCrawlers :: GetCrawlers

Creates a value of GetCrawlers with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gNextToken - A continuation token, if this is a continuation request.

gMaxResults - The number of Crawlers to return on each call.

data GetCrawlers

See: getCrawlers smart constructor.

Request Lenses

gNextToken :: Lens' GetCrawlers (Maybe Text)

A continuation token, if this is a continuation request.

gMaxResults :: Lens' GetCrawlers (Maybe Natural)

The number of Crawlers to return on each call.

Destructuring the Response

getCrawlersResponse

Arguments

:: Int
grsResponseStatus

-> GetCrawlersResponse

Creates a value of GetCrawlersResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

grsNextToken - A continuation token, if the returned list has not reached the end of those defined in this customer account.

grsCrawlers - A list of Crawler metadata.

grsResponseStatus - -- | The response status code.

data GetCrawlersResponse

See: getCrawlersResponse smart constructor.

Response Lenses

grsNextToken :: Lens' GetCrawlersResponse (Maybe Text)

A continuation token, if the returned list has not reached the end of those defined in this customer account.

grsCrawlers :: Lens' GetCrawlersResponse [Crawler]

A list of Crawler metadata.

grsResponseStatus :: Lens' GetCrawlersResponse Int

- | The response status code.