Network.AWS.Glue.GetDatabases

Description

Retrieves all Databases defined in a given Data Catalog.

Creating a Request

getDatabases :: GetDatabases

Creates a value of GetDatabases with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdCatalogId - The ID of the Data Catalog from which to retrieve Databases . If none is supplied, the AWS account ID is used by default.

gdNextToken - A continuation token, if this is a continuation call.

gdMaxResults - The maximum number of databases to return in one response.

data GetDatabases

See: getDatabases smart constructor.

Request Lenses

gdCatalogId :: Lens' GetDatabases (Maybe Text)

The ID of the Data Catalog from which to retrieve Databases . If none is supplied, the AWS account ID is used by default.

gdNextToken :: Lens' GetDatabases (Maybe Text)

A continuation token, if this is a continuation call.

gdMaxResults :: Lens' GetDatabases (Maybe Natural)

The maximum number of databases to return in one response.

Destructuring the Response

getDatabasesResponse

Arguments

:: Int
gdsrsResponseStatus

-> GetDatabasesResponse

Creates a value of GetDatabasesResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gdsrsNextToken - A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

gdsrsResponseStatus - -- | The response status code.

gdsrsDatabaseList - A list of Database objects from the specified catalog.

data GetDatabasesResponse

See: getDatabasesResponse smart constructor.

Response Lenses

gdsrsNextToken :: Lens' GetDatabasesResponse (Maybe Text)

A continuation token for paginating the returned list of tokens, returned if the current segment of the list is not the last.

gdsrsResponseStatus :: Lens' GetDatabasesResponse Int

- | The response status code.

gdsrsDatabaseList :: Lens' GetDatabasesResponse [Database]

A list of Database objects from the specified catalog.