Network.AWS.Glue.GetPartition

Description

Retrieves information about a specified partition.

Creating a Request

getPartition

Arguments

:: Text
gpDatabaseName

-> Text
gpTableName

-> GetPartition

Creates a value of GetPartition with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gpCatalogId - The ID of the Data Catalog where the partition in question resides. If none is supplied, the AWS account ID is used by default.

gpDatabaseName - The name of the catalog database where the partition resides.

gpTableName - The name of the partition's table.

gpPartitionValues - The values that define the partition.

data GetPartition

See: getPartition smart constructor.

Request Lenses

gpCatalogId :: Lens' GetPartition (Maybe Text)

The ID of the Data Catalog where the partition in question resides. If none is supplied, the AWS account ID is used by default.

gpDatabaseName :: Lens' GetPartition Text

The name of the catalog database where the partition resides.

gpTableName :: Lens' GetPartition Text

The name of the partition's table.

gpPartitionValues :: Lens' GetPartition [Text]

The values that define the partition.

Destructuring the Response

getPartitionResponse

Arguments

:: Int
gprsResponseStatus

-> GetPartitionResponse

Creates a value of GetPartitionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gprsPartition - The requested information, in the form of a Partition object.

gprsResponseStatus - -- | The response status code.

data GetPartitionResponse

See: getPartitionResponse smart constructor.

Response Lenses

gprsPartition :: Lens' GetPartitionResponse (Maybe Partition)

The requested information, in the form of a Partition object.

gprsResponseStatus :: Lens' GetPartitionResponse Int

- | The response status code.