Network.AWS.Glue.GetTableVersions

Description

Retrieves a list of strings that identify available versions of a specified table.

Creating a Request

getTableVersions

Arguments

:: Text
gtvDatabaseName

-> Text
gtvTableName

-> GetTableVersions

Creates a value of GetTableVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtvCatalogId - The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.

gtvNextToken - A continuation token, if this is not the first call.

gtvMaxResults - The maximum number of table versions to return in one response.

gtvDatabaseName - The database in the catalog in which the table resides.

gtvTableName - The name of the table.

data GetTableVersions

See: getTableVersions smart constructor.

Request Lenses

gtvCatalogId :: Lens' GetTableVersions (Maybe Text)

The ID of the Data Catalog where the tables reside. If none is supplied, the AWS account ID is used by default.

gtvNextToken :: Lens' GetTableVersions (Maybe Text)

A continuation token, if this is not the first call.

gtvMaxResults :: Lens' GetTableVersions (Maybe Natural)

The maximum number of table versions to return in one response.

gtvDatabaseName :: Lens' GetTableVersions Text

The database in the catalog in which the table resides.

gtvTableName :: Lens' GetTableVersions Text

The name of the table.

Destructuring the Response

getTableVersionsResponse

Arguments

:: Int
gtvrsResponseStatus

-> GetTableVersionsResponse

Creates a value of GetTableVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtvrsTableVersions - A list of strings identifying available versions of the specified table.

gtvrsNextToken - A continuation token, if the list of available versions does not include the last one.

gtvrsResponseStatus - -- | The response status code.

data GetTableVersionsResponse

See: getTableVersionsResponse smart constructor.

Response Lenses

gtvrsTableVersions :: Lens' GetTableVersionsResponse [TableVersion]

A list of strings identifying available versions of the specified table.

gtvrsNextToken :: Lens' GetTableVersionsResponse (Maybe Text)

A continuation token, if the list of available versions does not include the last one.

gtvrsResponseStatus :: Lens' GetTableVersionsResponse Int

- | The response status code.