Network.AWS.Glue.GetTrigger

Description

Retrieves the definition of a trigger.

Creating a Request

getTrigger

Arguments

:: Text
gtName

-> GetTrigger

Creates a value of GetTrigger with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtName - The name of the trigger to retrieve.

data GetTrigger

See: getTrigger smart constructor.

Request Lenses

gtName :: Lens' GetTrigger Text

The name of the trigger to retrieve.

Destructuring the Response

getTriggerResponse

Arguments

:: Int
gtrsResponseStatus

-> GetTriggerResponse

Creates a value of GetTriggerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gtrsTrigger - The requested trigger definition.

gtrsResponseStatus - -- | The response status code.

data GetTriggerResponse

See: getTriggerResponse smart constructor.

Response Lenses

gtrsTrigger :: Lens' GetTriggerResponse (Maybe Trigger)

The requested trigger definition.

gtrsResponseStatus :: Lens' GetTriggerResponse Int

- | The response status code.