Network.AWS.Glue.GetUserDefinedFunctions

Description

Retrieves a multiple function definitions from the Data Catalog.

Creating a Request

getUserDefinedFunctions

Arguments

:: Text
gudfDatabaseName

-> Text
gudfPattern

-> GetUserDefinedFunctions

Creates a value of GetUserDefinedFunctions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gudfCatalogId - The ID of the Data Catalog where the functions to be retrieved are located. If none is supplied, the AWS account ID is used by default.

gudfNextToken - A continuation token, if this is a continuation call.

gudfMaxResults - The maximum number of functions to return in one response.

gudfDatabaseName - The name of the catalog database where the functions are located.

gudfPattern - An optional function-name pattern string that filters the function definitions returned.

data GetUserDefinedFunctions

See: getUserDefinedFunctions smart constructor.

Request Lenses

gudfCatalogId :: Lens' GetUserDefinedFunctions (Maybe Text)

The ID of the Data Catalog where the functions to be retrieved are located. If none is supplied, the AWS account ID is used by default.

gudfNextToken :: Lens' GetUserDefinedFunctions (Maybe Text)

A continuation token, if this is a continuation call.

gudfMaxResults :: Lens' GetUserDefinedFunctions (Maybe Natural)

The maximum number of functions to return in one response.

gudfDatabaseName :: Lens' GetUserDefinedFunctions Text

The name of the catalog database where the functions are located.

gudfPattern :: Lens' GetUserDefinedFunctions Text

An optional function-name pattern string that filters the function definitions returned.

Destructuring the Response

getUserDefinedFunctionsResponse

Arguments

:: Int
gudfrsResponseStatus

-> GetUserDefinedFunctionsResponse

Creates a value of GetUserDefinedFunctionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gudfrsNextToken - A continuation token, if the list of functions returned does not include the last requested function.

gudfrsUserDefinedFunctions - A list of requested function definitions.

gudfrsResponseStatus - -- | The response status code.

data GetUserDefinedFunctionsResponse

See: getUserDefinedFunctionsResponse smart constructor.

Response Lenses

gudfrsNextToken :: Lens' GetUserDefinedFunctionsResponse (Maybe Text)

A continuation token, if the list of functions returned does not include the last requested function.

gudfrsUserDefinedFunctions :: Lens' GetUserDefinedFunctionsResponse [UserDefinedFunction]

A list of requested function definitions.

gudfrsResponseStatus :: Lens' GetUserDefinedFunctionsResponse Int

- | The response status code.