Network.AWS.Glue.StopCrawler

Description

If the specified Crawler is running, stops the crawl.

Creating a Request

stopCrawler

Arguments

:: Text
sName

-> StopCrawler

Creates a value of StopCrawler with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

sName - Name of the Crawler to stop.

data StopCrawler

See: stopCrawler smart constructor.

Request Lenses

sName :: Lens' StopCrawler Text

Name of the Crawler to stop.

Destructuring the Response

stopCrawlerResponse

Arguments

:: Int
srsResponseStatus

-> StopCrawlerResponse

Creates a value of StopCrawlerResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

srsResponseStatus - -- | The response status code.

data StopCrawlerResponse

See: stopCrawlerResponse smart constructor.

Response Lenses

srsResponseStatus :: Lens' StopCrawlerResponse Int

- | The response status code.