Network.AWS.Glue.UpdateDatabase

Description

Updates an existing database definition in a Data Catalog.

Creating a Request

updateDatabase

Arguments

:: Text
udName

-> DatabaseInput
udDatabaseInput

-> UpdateDatabase

Creates a value of UpdateDatabase with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udCatalogId - The ID of the Data Catalog in which the metadata database resides. If none is supplied, the AWS account ID is used by default.

udName - The name of the metadata database to update in the catalog.

udDatabaseInput - A DatabaseInput object specifying the new definition of the metadata database in the catalog.

data UpdateDatabase

See: updateDatabase smart constructor.

Request Lenses

udCatalogId :: Lens' UpdateDatabase (Maybe Text)

The ID of the Data Catalog in which the metadata database resides. If none is supplied, the AWS account ID is used by default.

udName :: Lens' UpdateDatabase Text

The name of the metadata database to update in the catalog.

udDatabaseInput :: Lens' UpdateDatabase DatabaseInput

A DatabaseInput object specifying the new definition of the metadata database in the catalog.

Destructuring the Response

updateDatabaseResponse

Arguments

:: Int
udrsResponseStatus

-> UpdateDatabaseResponse

Creates a value of UpdateDatabaseResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

udrsResponseStatus - -- | The response status code.

data UpdateDatabaseResponse

See: updateDatabaseResponse smart constructor.

Response Lenses

udrsResponseStatus :: Lens' UpdateDatabaseResponse Int

- | The response status code.