Network.AWS.Glue.UpdateTable

Description

Updates a metadata table in the Data Catalog.

Creating a Request

updateTable

Arguments

:: Text
utDatabaseName

-> TableInput
utTableInput

-> UpdateTable

Creates a value of UpdateTable with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utCatalogId - The ID of the Data Catalog where the table resides. If none is supplied, the AWS account ID is used by default.

utDatabaseName - The name of the catalog database in which the table resides.

utTableInput - An updated TableInput object to define the metadata table in the catalog.

data UpdateTable

See: updateTable smart constructor.

Request Lenses

utCatalogId :: Lens' UpdateTable (Maybe Text)

The ID of the Data Catalog where the table resides. If none is supplied, the AWS account ID is used by default.

utDatabaseName :: Lens' UpdateTable Text

The name of the catalog database in which the table resides.

utTableInput :: Lens' UpdateTable TableInput

An updated TableInput object to define the metadata table in the catalog.

Destructuring the Response

updateTableResponse

Arguments

:: Int
utrsResponseStatus

-> UpdateTableResponse

Creates a value of UpdateTableResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

utrsResponseStatus - -- | The response status code.

data UpdateTableResponse

See: updateTableResponse smart constructor.

Response Lenses

utrsResponseStatus :: Lens' UpdateTableResponse Int

- | The response status code.