Network.AWS.Glue.UpdateUserDefinedFunction

Description

Updates an existing function definition in the Data Catalog.

Creating a Request

updateUserDefinedFunction

Arguments

:: Text
uudfDatabaseName

-> Text
uudfFunctionName

-> UserDefinedFunctionInput
uudfFunctionInput

-> UpdateUserDefinedFunction

Creates a value of UpdateUserDefinedFunction with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uudfCatalogId - The ID of the Data Catalog where the function to be updated is located. If none is supplied, the AWS account ID is used by default.

uudfDatabaseName - The name of the catalog database where the function to be updated is located.

uudfFunctionName - The name of the function.

uudfFunctionInput - A FunctionInput object that re-defines the function in the Data Catalog.

data UpdateUserDefinedFunction

See: updateUserDefinedFunction smart constructor.

Request Lenses

uudfCatalogId :: Lens' UpdateUserDefinedFunction (Maybe Text)

The ID of the Data Catalog where the function to be updated is located. If none is supplied, the AWS account ID is used by default.

uudfDatabaseName :: Lens' UpdateUserDefinedFunction Text

The name of the catalog database where the function to be updated is located.

uudfFunctionName :: Lens' UpdateUserDefinedFunction Text

The name of the function.

uudfFunctionInput :: Lens' UpdateUserDefinedFunction UserDefinedFunctionInput

A FunctionInput object that re-defines the function in the Data Catalog.

Destructuring the Response

updateUserDefinedFunctionResponse

Arguments

:: Int
uudfrsResponseStatus

-> UpdateUserDefinedFunctionResponse

Creates a value of UpdateUserDefinedFunctionResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uudfrsResponseStatus - -- | The response status code.

data UpdateUserDefinedFunctionResponse

See: updateUserDefinedFunctionResponse smart constructor.

Response Lenses

uudfrsResponseStatus :: Lens' UpdateUserDefinedFunctionResponse Int

- | The response status code.